ReactElite WorkFlow
ReactElite Workflow is a continuous integration and delivery system that automates software development processes for ReactElite projects. It streamlines the process of installing dependencies, building the project, running tests, and deploying the project to a hosting service. The workflow is customizable and allows for triggering notifications or other automated actions based on the outcome of the workflow. This makes the development process more efficient, reliable, and consistent. The workflow is defined in a YAML file, which can be easily modified to meet the specific needs of a project.